+++ src/main/kotlin/com/itmo/microservices/demo/orders/api/controller/OrderController.kt
@@ -2,13 +2,9 @@ package com.itmo.microservices.demo.orders.api.controller
 
 import com.itmo.microservices.demo.delivery.api.model.DeliveryModel
 import com.itmo.microservices.demo.delivery.api.service.DeliveryService
-import com.itmo.microservices.demo.orders.api.model.OrderModel
-import com.itmo.microservices.demo.orders.api.model.PaymentModel
 import com.itmo.microservices.demo.orders.api.service.OrderService
-import com.itmo.microservices.demo.shoppingCartService.api.dto.ShoppingCartDTO
 import com.itmo.microservices.demo.shoppingCartService.api.service.CartService
 import io.swagger.v3.oas.annotations.Operation
-import io.swagger.v3.oas.annotations.Parameter
 import io.swagger.v3.oas.annotations.media.Content
 import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.security.SecurityRequirement
@@ -16,7 +12,6 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.web.bind.annotation.*
 import java.util.*
-import javax.transaction.NotSupportedException
 
 @RestController
 class OrderController(private val orderService: OrderService,
