+++ src/main/kotlin/com/itmo/microservices/demo/shoppingCartService/impl/entity/ShoppingCart.kt
@@ -1,8 +1,6 @@
 package com.itmo.microservices.demo.shoppingCartService.impl.entity
 
 
-import org.hibernate.annotations.GenericGenerator
-import org.hibernate.annotations.Type
 import java.util.*
 import javax.persistence.*
 
@@ -10,12 +8,8 @@ import javax.persistence.*
 @Table(name="Cart304")
 class ShoppingCart {
     @Id
-    @Type(type = "uuid-char")
-    @GenericGenerator(
-            name = "UUID",
-            strategy = "org.hibernate.id.UUIDGenerator"
-    )
-    var id : UUID = UUID.randomUUID()
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    lateinit var id : UUID
     var status: String = "ACTIVE"
